Cleanup/fix range and dimension related traits/methods (#196)
axes_types was erroneously assuming that the the parent axis always had an offfset of 1
known_first and known_last work for CartesianIndices now
defined contiguous_axis, stride_rank, and offset1 for StrideIndex
has_dimnames returns StaticBool now. Not breaking because no downstream dependencies yet.

+++ src/array_index.jl
@@ -194,7 +194,7 @@ struct StrideIndex{N,R,C,S,O} <: ArrayIndex{N}
     offsets::O
 
     function StrideIndex{N,R,C}(s::S, o::O) where {N,R,C,S,O}
-        return new{N,R::NTuple{N,Int},C::Int,S,O}(s, o)
+        return new{N,R::NTuple{N,Int},C,S,O}(s, o)
     end
     function StrideIndex{N,R,C}(a::A) where {N,R,C,A}
         return StrideIndex{N,R,C}(strides(a), offsets(a))
